Add files via upload
This GitHub repository contains a sentiment analysis classifier that is trained on Amazon customer reviews. The classification is done using two different models, namely NLTK's VADER and Hugging Face's Roberta transformers. The purpose of this project is to compare the performance of both models and present the results in a pairplot generated using the Seaborn library.

The sentiment analysis classifier is designed to automatically analyze customer reviews and determine whether they are positive, negative or neutral. The use of two different models provides a comprehensive analysis of the effectiveness of each model and their ability to classify reviews accurately.

The repository includes code and documentation for each step of the project, including data preprocessing, model training and evaluation, and visualization of the results. Overall, this project provides valuable insights into the effectiveness of different sentiment analysis models and can be useful for anyone interested in analyzing customer reviews or sentiment analysis in general.